Work started on JS card generation. Minor refactoring
Refactored cardOptions.js into cardOptions.php. Work started on #7 Still needs  to generate appendable properties.
